+++ tls/CHANGES.md
@@ -3,4 +3,4 @@
 - update `rustls` to `0.23.0`
 
 ## Remove
-- `AsyncRead` and `AsyncWrite` trait impl has been removed from `rustls::TlsStream`. Please use `xitca_io::io::AsyncReadWrite` adaptor as alternative when these traits are needed for `TlsStream`.
+- `AsyncRead` and `AsyncWrite` traits impl has been removed from `rustls::TlsStream`. Please use `xitca_io::io::PollIoAdapter` as alternative when these traits are needed.